requirements.txt
future
emoji
lxml
wget
beautifulsoup4
requests
python-telegram-bot==12.8
youtube_search_python
sqlalchemy==1.4.11
PyDictionary
psycopg2-binary
feedparser
hachoir
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow
CurrencyConverter
jikanpy
speedtest-cli
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon
telegraph
heroku3
nekos.py
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
TgCrypto
Pyrogram==1.4.16
youtube_search
asyncio
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
opencv-python-headless
bing_image_downloader
search_engine_parser
pytz
pykeyboard
aiofiles
motor
telegram
gpytranslate
yt_dlp
gtts
python-arq
envparse
